+++ src/test/java/net/imglib2/algorithm/kdtree/SplitHyperPlaneKDTreeExample.java
@@ -31,8 +31,8 @@ public class SplitHyperPlaneKDTreeExample
 		final ArrayList< Point > points = new ArrayList< Point >();
 		for ( int i = 0; i < nPoints; ++i )
 		{
-			final int x = rand.nextInt( w );
-			final int y = rand.nextInt( h );
+			final long x = rand.nextInt( w );
+			final long y = rand.nextInt( h );
 			points.add( new Point( x, y ) );
 		}
